Smashed PNG Icons: Convert PNG icons into a a dual Workbench/PNG format
This makes .info icons that are compatible with KS 1.3 ... 3.9

Encoded in the file are 4-color AOS 1.3 style icons, 16-color AOS 3.5
icons (using the Scalos pallette), and PNG icons for AROS use.

Hopefully we can get Morphos/AmigaOS on board with using this format.

NOTE: AmigaOS 3.9's icon editor can edit these icons, but it will strip out
      the PNG data, instead of preserving it as the IFF spec specifies.

      Bad AmigaOS! No cookie!


git-svn-id: https://svn.aros.org/svn/aros/trunk@41726 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/icon/deletediskobject.c
@@ -6,6 +6,7 @@
     Lang: english
 */
 #include <proto/dos.h>
+#include <proto/workbench.h>
 #include "icon_intern.h"
 
 /*****************************************************************************
@@ -34,7 +35,6 @@
     EXAMPLE
 
     BUGS
-	Does not yet notify workbench about the deletion.
 
     SEE ALSO
 
@@ -46,7 +46,11 @@
 {
     AROS_LIBFUNC_INIT
     UBYTE * infofilename;
-    BOOL    success;
+    BOOL    success = FALSE;
+    BPTR lock, parent;
+
+    if (!WorkbenchBase)
+        WorkbenchBase = TaggedOpenLibrary(TAGGEDOPEN_WORKBENCH);
 
     if (!(infofilename = (UBYTE*)AllocVec (strlen(name) + 6,
 	MEMF_ANY | MEMF_CLEAR)
@@ -57,7 +61,18 @@
     strcpy (infofilename, name);
     strcat (infofilename, ".info");
 
-    success = DeleteFile (infofilename);
+    lock = Lock(infofilename, SHARED_LOCK);
+    if (lock) {
+        parent = ParentDir(lock);
+        if (parent) {
+            success = DeleteFile (infofilename);
+            if (success && WorkbenchBase) {
+                UpdateWorkbench(FilePart(name), parent, UPDATEWB_ObjectRemoved);
+            }
+            UnLock(parent);
+        }
+        UnLock(lock);
+    }
 
     FreeVec (infofilename);
